Move representation chooser into own file
Will make testing AdaptiveStream possible
